requirements.txt
streamlit==1.32.0
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
poppler-utils==0.1.0 